+++ examples/sink_source.jl
@@ -44,8 +44,18 @@ function generate_data()
         Direct(12, nodes[1], nodes[2], Linear())
     ]
 
+    # Variables for the individual entries of the time structure
+    op_duration = 2 # Each operational period has a duration of 2
+    op_number = 4   # There are in total 4 operational periods
+    operational_periods = SimpleTimes(op_number, op_duration)
+
+    # The number of operational periods times the duration of the operational periods, which
+    # can also be extracted using the function `duration` which corresponds to the total
+    # duration of the operational periods in a `SimpleTimes` structure
+    op_per_strat = duration(operational_periods)
+
     # Creation of the time structure and global data
-    T = TwoLevel(4, 1, SimpleTimes(4, 2))
+    T = TwoLevel(4, 1, operational_periods; op_per_strat)
     model = OperationalModel(Dict(CO2 => FixedProfile(10)), CO2)
 
     # WIP data structure